feat: rework for carousel of pictures
Made the carousel into a horizontal Scrollview
with square images, that expand when clicked on.
Arrows are styled and displayed only when the image is
in full screen mode. Modified the resizing/compressing
of the images to be a percentage of their original dimensions
instead of a fixed-size (needed to avoid distortion in full-mode).
Issue Details now calculates the format of each picture
(portrait/landscape) to update the dimensions accordingly
in full mode.