fix: display of elements in view vs scrollview
Transformed the View of contents into a ScrollView
that follows the size of the gray background instead
of the opposite way. Made the background to be a fixed size,
relative to the screen, to avoid display problems when
maintenance request is too long or too short. Made the contents
of the scrollview cut nicely when sliding on the rounded edges.