feat: CloudFunction for notification sending
Initalize Firestore Cloud Functions in the project.
Create a cloud function in index.ts of functions folder
to send secure notification when cycle is about to end.
Add method in firestore.ts to create a document when a
notification should be sent to a specific user.
Calculation of timer inside frontend calls this function
3 minutes before estimated end time.